1. Modified AI assistant via json configuration in basic.json.    The intention is to set up the assistant to take a call and gather the name, email of the client and then schudule a meeting by gathering the time, date, and timezone/location of the client.
2. Added Private key for VAPI dashboard in the .env file

3. Removed custom llm since the goal of this MVP is to highlight the intergration of schuduling meetings with googles API via a VAPI voice agent.

Next Actions:
I hope to avoid having to set up a database so this MVP will not store call data but pass a payload to be used and deleted or dropped.
This means I should be able to work on the googleCalendar function which I think I will rename to createGoogleMeeting.